Hooked up Kafka inventory topics to inventory adapter.
