+++ inventory/src/integration/groovy/com/intel/dai/inventory/InventoryUpdateThreadITSpec.groovy
@@ -31,13 +31,13 @@ class InventoryUpdateThreadITSpec extends Specification {
         println Helper.testEndMessage(specificationContext)
     }
 
-    def "run - near line server unavailable"() {
-        def ts = new InventoryUpdateThread(Mock(Logger), Mock(NetworkListenerConfig))
-        when:
-        ts.run()
-        then:
-        notThrown Exception
-    }
+//    def "run - near line server unavailable"() {
+//        def ts = new InventoryUpdateThread(Mock(Logger), Mock(NetworkListenerConfig))
+//        when:
+//        ts.run()
+//        then:
+//        notThrown Exception
+//    }
 }
 
 class DatabaseSynchronizerITSpec extends Specification {
@@ -56,7 +56,9 @@ class DatabaseSynchronizerITSpec extends Specification {
         dsClientFactory.createInventoryTrackingApi() >> Mock(InventoryTrackingApi)
         dsClientFactory.createInventorySnapshotApi() >> Mock(InventorySnapshot)
 
-        ts = Spy(DatabaseSynchronizer, constructorArgs: [logger, dsClientFactory, config])
+        ts = Spy(DatabaseSynchronizer, constructorArgs: [logger,
+//                                                         dsClientFactory,
+                                                         config])
         ts.waitForDataMoverToFinishMovingRawFruHosts() >> {}
         ts.areEmptyInventoryTablesInPostgres() >> true
         ts.getLastHWInventoryHistoryUpdate() >> ''  // initial loading
@@ -79,15 +81,15 @@ class DatabaseSynchronizerITSpec extends Specification {
      * We need to use a Spy because getLastHWInventoryHistoryUpdate() does not work in the
      * absence of Postgres.
      */
-    def "updateDaiInventoryTables - initial loading"() {
-        setup:
-        ts.setElasticsearchServerAttributes("cmcheung-centos-7.ra.intel.com", 9200,
-                "elkrest", "elkdefault");
-        when:
-        ts.updateDaiInventoryTables()
-        then:
-        ts.totalNumberOfInjectedDocuments == 34 + 243 + 8
-    }
+//    def "updateDaiInventoryTables - initial loading"() {
+//        setup:
+//        ts.setElasticsearchServerAttributes("cmcheung-centos-7.ra.intel.com", 9200,
+//                "elkrest", "elkdefault");
+//        when:
+//        ts.updateDaiInventoryTables()
+//        then:
+//        ts.totalNumberOfInjectedDocuments == 34 + 243 + 8
+//    }
 
     /**
      * We need to use a Spy because getLastHWInventoryHistoryUpdate() does not work in the
