+++ inventory/src/integration/resources/scripts/restart_voltdb.sh
@@ -2,7 +2,7 @@
 
 docker-compose -f docker-compose/db.yml down
 docker-compose -f docker-compose/db.yml up -d
-sleep 15
+sleep 25
 createdb --host=css-centos-8-00.ra.intel.com --username=postgres dai
 psql --host=css-centos-8-00.ra.intel.com --username=postgres --dbname=dai < postgres_inventory.sql
 psql --host=css-centos-8-00.ra.intel.com --username=postgres --dbname=dai < postgres_inventory_data.sql
