+++ inventory/src/main/java/com/intel/dai/inventory/NetworkListenerProviderForeignBus.java
@@ -11,6 +11,7 @@ import com.intel.dai.inventory.api.HWInvNotificationTranslator;
 import com.intel.dai.inventory.api.pojo.scn.ForeignHWInvChangeNotification;
 import com.intel.dai.network_listener.*;
 import com.intel.logging.Logger;
+import org.apache.commons.lang3.tuple.ImmutablePair;
 
 import java.time.Instant;
 import java.util.ArrayList;
@@ -33,51 +34,69 @@ public class NetworkListenerProviderForeignBus implements NetworkListenerProvide
      * that shares a common state update.  This translate into an array of CommonDataFormat entries, each
      * describing a single component. </p>
      *
-     * @param subject The subject or topic associated with this message.
-     * @param scnJson state change notification as a json string
+     * @param topic The topic or topic associated with this message.
+     * @param inventoryJson state change notification as a json string
      * @param config network listener configuration
      * @return an array of CommonDataFormat entries equivalent to scnJson
      * @throws NetworkListenerProviderException network listener provider exception
      */
     @Override
-    public List<CommonDataFormat> processRawStringData(String subject, String scnJson, NetworkListenerConfig config)
+    public List<CommonDataFormat> processRawStringData(String topic, String inventoryJson, NetworkListenerConfig config)
             throws NetworkListenerProviderException {
+        log_.debug("Received %s: %s", topic, inventoryJson);
 
-        long currentTimestamp = currentUTCInNanoseconds();
+        // First start with performing the voltdb update right here
+        // CMC_TODO: Refactor working code into actOnData; actually, it is not worth it
 
-        ForeignHWInvChangeNotification foreignInventoryChangeNotification =
-                new HWInvNotificationTranslator(log_).toPOJO(scnJson);
-        if (foreignInventoryChangeNotification == null) {
-            throw new NetworkListenerProviderException("Cannot extract foreignInventoryChangeNotification");
+        DatabaseSynchronizer synchronizer = new DatabaseSynchronizer(log_, config_);
+        switch (topic) {
+            case "kafka_dimm":
+                synchronizer.ingestRawDimm(new ImmutablePair<>(topic, inventoryJson));
+                break;
+            case "kafka_fru_host":
+                synchronizer.ingestRawFruHost(new ImmutablePair<>(topic, inventoryJson));
+                break;
+            default:
+                log_.error("Unexpected kafka topic: %s", topic);
+                break;
         }
 
-        BootState newComponentState = toBootState(foreignInventoryChangeNotification.State);
-        if (newComponentState == null) {
-            log_.info("HWI:%n  ignoring unsupported scn notification state: %s",
-                    foreignInventoryChangeNotification.State);
-            return new ArrayList<>();
-        }
-
+//        DatabaseSynchronizer ds = new DatabaseSynchronizer(log_, null, config);
+//        long currentTimestamp = currentUTCInNanoseconds();
+//
+//        ForeignHWInvChangeNotification foreignInventoryChangeNotification =
+//                new HWInvNotificationTranslator(log_).toPOJO(scnJson);
+//        if (foreignInventoryChangeNotification == null) {
+//            throw new NetworkListenerProviderException("Cannot extract foreignInventoryChangeNotification");
+//        }
+//
+//        BootState newComponentState = toBootState(foreignInventoryChangeNotification.State);
+//        if (newComponentState == null) {
+//            log_.info("HWI:%n  ignoring unsupported scn notification state: %s",
+//                    foreignInventoryChangeNotification.State);
+//            return new ArrayList<>();
+//        }
+//
         List<CommonDataFormat> workItems = new ArrayList<>();
-        for (String foreign: foreignInventoryChangeNotification.Components) {
-            String location;
-            try {
-                location = CommonFunctions.convertForeignToLocation(foreign);  // this fails if translation map is outdated
-            } catch(ConversionException e) {
-                location = foreign; // for debugging purpose
-//                throw new NetworkListenerProviderException(
-//                        "Failed to convert the foreign location to a DAI location", e);
-            }
-
-            CommonDataFormat workItem = new CommonDataFormat(
-                    currentTimestamp,
-                    location,
-                    DataType.InventoryChangeEvent);
-            workItem.setStateChangeEvent(newComponentState);
-            workItem.storeExtraData(FOREIGN_KEY, foreign);
-            workItems.add(workItem);
-        }
-        log_.info("HWI:%n  Extracted %d work items", workItems.size());
+//        for (String foreign: foreignInventoryChangeNotification.Components) {
+//            String location;
+//            try {
+//                location = CommonFunctions.convertForeignToLocation(foreign);  // this fails if translation map is outdated
+//            } catch(ConversionException e) {
+//                location = foreign; // for debugging purpose
+////                throw new NetworkListenerProviderException(
+////                        "Failed to convert the foreign location to a DAI location", e);
+//            }
+//
+//            CommonDataFormat workItem = new CommonDataFormat(
+//                    currentTimestamp,
+//                    location,
+//                    DataType.InventoryChangeEvent);
+//            workItem.setStateChangeEvent(newComponentState);
+//            workItem.storeExtraData(FOREIGN_KEY, foreign);
+//            workItems.add(workItem);
+//        }
+//        log_.info("HWI:%n  Extracted %d work items", workItems.size());
         return workItems;   // to be consumed by com.intel.dai.network_listener.processMessage()
     }
 
@@ -138,16 +157,19 @@ public class NetworkListenerProviderForeignBus implements NetworkListenerProvide
         if(config_ == null)
             getConfig(config);
 
-        BootState bootState = workItem.getStateEvent();
-        if (!isSupportedInventoryEvents(bootState)) {
-            log_.debug("Unsupported boot state=%d", bootState);
-            return;
-        }
+        // Enqueue inventory json into a voltdb table so that a dedicated thread can consume them.
+        // This means that inventory loading thread cannot shutdown by itself.  Use an interrupt
+        // to shut the thread down.
 
-        // CMC_TODO: The following will be removed
-        log_.info("Starting InventoryUpdateThread ...");
-        Thread t = new Thread(new InventoryUpdateThread(log_, config));
-        t.start();  // background update of inventory
+//        BootState bootState = workItem.getStateEvent();
+//        if (!isSupportedInventoryEvents(bootState)) {
+//            log_.debug("Unsupported boot state=%d", bootState);
+//            return;
+//        }
+//
+//        log_.info("Starting InventoryUpdateThread ...");
+//        Thread t = new Thread(new InventoryUpdateThread(log_, config));
+//        t.start();  // background update of inventory
     }
 
     private boolean isSupportedInventoryEvents(BootState bootState) {
