+++ inventory/src/main/java/com/intel/dai/inventory/api/HWInvDiscovery.java
@@ -131,7 +131,7 @@ public class HWInvDiscovery {
             throw new RESTClientException("HWI:%n  Was unable to find the configuration file: HWInvDiscoveryConfig.json");
 
         sess = toHWDiscoverySession(configPath);
-        log.debug("HWI:%n  toHWDiscoverySession(configPath=%s) => %s", configPath, sess.toString());
+        log.debug("toHWDiscoverySession(configPath=%s) => %s", configPath, sess.toString());
 
         RESTClient restClient = RESTClientFactory.getInstance("apache", log);
         if (restClient == null) {
