+++ inventory/src/test/groovy/com/intel/dai/inventory/InventoryUpdateThreadSpec.groovy
@@ -15,14 +15,16 @@ import com.intel.logging.Logger
 import spock.lang.Specification
 
 class DatabaseSynchronizerSpec extends Specification {
-    def ts = new DatabaseSynchronizer(Mock(Logger), Mock(DataStoreFactory), Mock(NetworkListenerConfig))
+    def ts = new DatabaseSynchronizer(Mock(Logger),
+//            Mock(DataStoreFactory),
+            Mock(NetworkListenerConfig))
 
-    def "initializeDependencies"() {
-        when: ts.initializeDependencies()
-        then:
-        ts.util_ != null
-        ts.foreignInventoryDatabaseClient_ != null
-    }
+//    def "initializeDependencies"() {
+//        when: ts.initializeDependencies()
+//        then:
+//        ts.util_ != null
+//        ts.foreignInventoryDatabaseClient_ != null
+//    }
 
     def "extractChangedNodeLocations -- empty input"() {
         expect:
@@ -125,12 +127,12 @@ class DatabaseSynchronizerSpec extends Specification {
         ts.ingestRawInventorySnapshot("", []) == 0
     }
 
-    def "updateDaiInventoryTables"() {
-        ts.util_ = new HWInvUtilImpl(Mock(Logger))
-        ts.foreignInventoryDatabaseClient_ = Mock(ForeignInventoryClient)
-        ts.foreignInventoryDatabaseClient_.getCanonicalHWInvHistoryJson(_) >> null
-
-        expect:
-        ts.updateDaiInventoryTables()
-    }
+//    def "updateDaiInventoryTables"() {
+//        ts.util_ = new HWInvUtilImpl(Mock(Logger))
+//        ts.foreignInventoryDatabaseClient_ = Mock(ForeignInventoryClient)
+//        ts.foreignInventoryDatabaseClient_.getCanonicalHWInvHistoryJson(_) >> null
+//
+//        expect:
+//        ts.updateDaiInventoryTables()
+//    }
 }
