+++ sandbox_pipelines/sandbox.components.pipeline.groovy
@@ -12,6 +12,9 @@ pipeline {
     }
     stages {
         stage('Restart Component Test Servers') {
+            environment {
+                PGPASSWORD = 'postgres'
+            }
             agent { label "DOCKER-TEST-SERVER-HOST" }
             steps {
                 script { utilities.fixFilesPermission() }
